requirements.txt
# requirements.txt
fastapi~=0.115.8
uvicorn[standard]~=0.34.0
sqlalchemy~=2.0.37
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1
google-generativeai
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.5.0
python-docx~=1.2.0
pypdf~=5.8.0
email-validator
dnspython~=2.7.0
idna~=3.10
cryptography~=45.0.5
anyio~=4.8.0
h11~=0.14.0
pip~=23.2.1
google~=3.0.0
protobuf~=5.29.5
pillow~=11.3.0
rsa~=4.9.1
pyasn1~=0.6.1
cffi~=1.17.1
typing_extensions~=4.12.2
lxml~=6.0.0
PyMuPDF~=1.26.3
grpcio~=1.73.1
ecdsa~=0.19.1
beautifulsoup4~=4.13.4
requests~=2.32.4
colorama~=0.4.6
sniffio~=1.3.1
six~=1.17.0
proto-plus~=1.26.1
click~=8.1.8
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.7.14
urllib3~=2.5.0
google-api-python-client~=2.176.0
grpcio-status~=1.71.2
pydantic~=2.10.6
httplib2~=0.22.0
tqdm~=4.67.1
starlette~=0.45.3
email_validator~=2.2.0
pydantic_core~=2.27.2
greenlet~=3.1.1
pyparsing~=3.2.3
psycopg2~=2.9.10
annotated-types~=0.7.0
charset-normalizer~=3.4.2
pycparser~=2.22
uritemplate~=4.2.0
bcrypt~=4.3.0
google-auth-httplib2~=0.2.0
python-multipart~=0.0.20
dotenv~=0.9.9
fpdf2~=2.8.3
defusedxml~=0.7.1
fonttools~=4.58.5
sympy~=1.13.3
jinja2
weasyprint
reportlab
gunicorn